requirements.txt
annotated-doc
annotated-types
anyio==4.12.1
bcrypt
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ddgs==9.10.0
duckduckgo-search>=6.0.0
fake-useragent==2.2.0
fastapi
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint-sqlite
lxml==6.0.2
numpy
openai
pandas
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.1.0
primp==0.15.0
pycparser==2.23
pydantic
pyjwt
pymysql
pypdfium2==5.2.0
python-dotenv
python-multipart
socksio==1.0.0
sqlalchemy
uvicorn